backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo App Backend API"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "django-filter>=23.0",
    "django-cors-headers>=4.3.0",
    "django-environ>=0.11.0",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.2",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
